src/bugbug/data/requirements_frozen.txt
aioamqp==0.12.0
aiohttp==3.4.4
asn1crypto==0.24.0
async-timeout==3.0.1
atomicwrites==1.2.1
attrs==18.2.0
boto3==1.9.71
botocore==1.12.71
bugbug==0.0.1
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.2
coveralls==1.5.1
cryptography==2.4.2
cymem==2.0.2
cytoolz==0.9.0.1
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
elasticsearch==6.3.1
en-core-web-sm==2.0.0
fancycompleter==0.8
flake8==3.6.0
flake8-coding==1.3.1
flake8-copyright==0.2.2
flake8-debugger==3.1.0
flake8-isort==2.6.0
flake8-mypy==17.8.0
flake8-quotes==1.0.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
icalendar==4.0.3
idna==2.7
imbalanced-learn==0.4.3
isort==4.3.4
jmespath==0.9.3
libmozdata==0.1.44
Logbook==1.4.1
mccabe==0.6.1
mohawk==0.3.4
more-itertools==4.3.0
mozdef-client==1.0.11
mozilla-cli-common==1.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.5.2
murmurhash==1.0.1
mypy==0.650
mypy-extensions==0.4.1
numpy==1.15.4
oauth2client==4.1.3
pdbpp==0.9.3
plac==0.9.6
pluggy==0.8.0
preshed==2.0.1
py==1.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
Pygments==2.3.1
pyOpenSSL==18.0.0
pytest==4.0.2
pytest-cov==2.6.0
pytest-runner==4.2
python-dateutil==2.6.1
python-hglib==2.6.1
pytz==2018.7
raven==6.10.0
regex==2018.1.10
requests==2.21.0
requests-futures==0.9.9
responses==0.10.5
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.2
scipy==1.2.0
setuptools-scm==3.1.0
six==1.12.0
slugid==1.0.7
spacy==2.0.18
structlog==18.2.0
taskcluster==6.0.0
taskcluster-urls==11.0.0
testfixtures==6.4.0
thinc==6.12.1
toolz==0.9.0
tqdm==4.28.1
typed-ast==1.1.1
ujson==1.35
uritemplate==3.0.0
urllib3==1.24.1
whatthepatch==0.0.5
wmctrl==0.3
wrapt==1.10.11
xgboost==0.81
yarl==1.3.0
